backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
pymongo==4.6.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
motor==3.3.2  # Async MongoDB driver